moved maker and pushover to plugins; initial maker setup docs
