+++ lib/bootevent.js
@@ -17,9 +17,11 @@ function boot (env) {
     // api and json object variables
     ///////////////////////////////////////////////////
     ctx.plugins = require('./plugins')().registerServerDefaults().init(env);
-    ctx.pushover = require('./pushover')(env);
-    ctx.maker = require('./maker')(env);
+
+    ctx.pushover = require('./plugins/pushover')(env);
+    ctx.maker = require('./plugins/maker')(env);
     ctx.pushnotify = require('./pushnotify')(env, ctx);
+
     ctx.entries = require('./entries')(env, ctx);
     ctx.treatments = require('./treatments')(env, ctx);
     ctx.devicestatus = require('./devicestatus')(env.devicestatus_collection, ctx);
