+++ lib/plugins/maker.js
